# feat: Docker container optimization for production deployment (#307)