… infra/database/lambda/db-init-handler.js