@krishagel 2025-10-04
## ✅ Review Feedback Addressed

Thank you for the thorough reviews! I've addressed all the critical and important feedback from the automated reviews.

### 🔴 Critical Security Fixes

**1. Read-Only Filesystem Implementation** (infra/lib/constructs/ecs-service.ts:277)
- ✅ Changed `readonlyRootFilesystem: false` → `true`
- ✅ Implemented tmpfs mounts for writable directories:
  - `/tmp` → 512MB
  - `/app/.next/cache` → 256MB
- ✅ Follows least privilege security principle while maintaining Next.js functionality

**2. Npm Dependency Installation Strategy** (Dockerfile:16)
- ✅ Removed `--omit=dev` flag from deps stage
- ✅ Ensures dev dependencies available for builder stage
- ✅ Prevents build failures from missing development packages

### 🟡 Important Improvements

**3. Logging Standards Compliance** (app/api/healthz/route.ts)
- ✅ Added proper error handling per CLAUDE.md standards
- ✅ Implemented try/catch with `createLogger` usage
- ✅ Returns 503 Service Unavailable on health check failures

**4. Configuration Portability** (Dockerfile + docker-compose.yml)
- ✅ Moved `NODE_OPTIONS` from Dockerfile to docker-compose.yml
- ✅ Improves Docker image portability across environments
- ✅ Maintains 70% memory allocation strategy (1400MB for 2GB container)

### ✅ Quality Checks

- ✅ `npm run lint` - PASSING (no warnings or errors)
- ✅ `npm run typecheck` - PASSING
- ✅ All original performance optimizations maintained:
  - BuildKit cache mounts (50-90% faster builds)
  - Tini for graceful shutdowns (<5s)
  - Lightweight /api/healthz endpoint
  - Node 22 Alpine (33% smaller image)

### 📊 Summary

All blocking and important issues have been resolved. The changes maintain the original performance benefits while significantly improving security posture and code quality.

**Ready for re-review!** 🚀